[flexible-ipam] Bridging mode and add Pod flows
This commit enables bridging mode for flexible-IPAM.
Currently K8s Node will route Pod traffic between interface
antrea-gw0 and uplink.
When enable flexible-IPAM, uplink will be bridged to Antrea managed
OVS bridge, which called "bridging mode".

+++ pkg/ovs/openflow/ofctrl_action.go
@@ -24,8 +24,8 @@ func (a *ofFlowAction) Drop() FlowBuilder {
 }
 
 // Output is an action to output packets to the specified ofport.
-func (a *ofFlowAction) Output(port int) FlowBuilder {
-	outputAction := ofctrl.NewOutputPort(uint32(port))
+func (a *ofFlowAction) Output(port uint32) FlowBuilder {
+	outputAction := ofctrl.NewOutputPort(port)
 	a.builder.ApplyAction(outputAction)
 	return a.builder
 }
